添加了方块的放置和破坏，重写了生物和方块的模型，添加了射线检测算法，添加了指针。

+++ src/com/sunrisestudio/util/math/HitResult.java
@@ -3,7 +3,7 @@ package com.sunrisestudio.util.math;
 import com.sunrisestudio.cubecraft.world.IWorldAccess;
 import com.sunrisestudio.cubecraft.world.entity.Entity;
 
-public record HitResult(HitBox aabb, int facing) {
+public record HitResult(HitBox aabb, byte facing) {
     public void hit(IWorldAccess worldAccess, Entity from) {
         this.aabb.getObject().onHit(from,worldAccess,this);
     }
